hass_host_sensors/requirements.txt
paho-mqtt==1.6.1
psutil==5.8.0